trigger add note cubit

+++ lib/models/note.dart
@@ -11,7 +11,7 @@ class Note extends HiveObject {
   @HiveField(2)
   final String date;
   @HiveField(3)
-  final int color;
+  final double color;
 
   Note({
     required this.title,
